# openapi.yaml
openapi: 3.0.3
info:
  title: NOC Chamados API
  description: |
    API REST para sistema de gestão de chamados de NOC (Network Operations Center).
    
    ## Autenticação
    A API usa autenticação JWT via HTTPOnly Cookie. Faça login primeiro usando `/auth/login`.
    
    ## Rate Limiting
    - Login: 5 tentativas por 15 minutos (por IP)
    
    ## Recursos
    - [Documentação Interativa](/api-docs)
    - [Guia Swagger](./SWAGGER_GUIDE.md)
    - [Coleção Postman](./NOC_Chamados.postman_collection.json)
    
  version: 1.2.3
  contact:
    name: Suporte NOC Chamados
  license:
    name: Proprietário
servers:
  - url: https://chamados.xmov.com.br/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento Local
...
paths:
...
  /links:
    get:
      tags:
        - Links
      summary: Listar Links
      description: Lista todos os links com filtros opcionais
      operationId: listarLinks
      security:
        - cookieAuth: []
      parameters:
        - name: ativo
          in: query
          schema:
            type: boolean
        - name: operadoraId
          in: query
          schema:
            type: integer
        - name: popId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Link'

    post:
      tags:
        - Links
      summary: Criar Link
      description: Cria um novo link
      operationId: criarLink
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarLink'
      responses:
        '201':
          description: Link criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
  /links/{id}:
    get:
      tags:
        - Links
      summary: Obter Link
      operationId: obterLink
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes do link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'

    put:
      tags:
        - Links
      summary: Atualizar Link
      operationId: atualizarLink
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarLink'
      responses:
        '200':
          description: Link atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'

    delete:
      tags:
        - Links
      summary: Deletar Link
      description: Exclui permanentemente um link arquivado (apenas supervisores)
      operationId: deletarLink
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Link deletado
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==============================================
  # TRANSPORTES
  # ==============================================
...
  /operadoras:
    get:
      tags:
        - Operadoras
      summary: Listar Operadoras
      operationId: listarOperadoras
      security:
        - cookieAuth: []
      parameters:
        - name: ativo
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de operadoras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Operadora'

    post:
      tags:
        - Operadoras
      summary: Criar Operadora
      operationId: criarOperadora
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarOperadora'
      responses:
        '201':
          description: Operadora criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operadora'
...
  /pops:
    get:
      tags:
        - POPs
      summary: Listar POPs
      operationId: listarPops
      security:
        - cookieAuth: []
      parameters:
        - name: ativo
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de POPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pop'

    post:
      tags:
        - POPs
      summary: Criar POP
      operationId: criarPop
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarPop'
      responses:
        '201':
          description: POP criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pop'
  /pops/{id}:
    get:
      tags:
        - POPs
      summary: Obter POP
      operationId: obterPop
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes do POP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pop'

    put:
      tags:
        - POPs
      summary: Atualizar POP
      operationId: atualizarPop
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarPop'
      responses:
        '200':
          description: POP atualizado

    delete:
      tags:
        - POPs
      summary: Deletar POP
      operationId: deletarPop
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: POP deletado

  # ==============================================
  # USUÁRIOS
  # ==============================================
...
  /tipos-falha:
    get:
      tags:
        - Tipos de Falha
      summary: Listar Tipos de Falha
      operationId: listarTiposFalha
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Lista de tipos de falha
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TipoFalha'

    post:
      tags:
        - Tipos de Falha
      summary: Criar Tipo de Falha
      operationId: criarTipoFalha
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
              properties:
                nome:
                  type: string
                  example: Fibra Rompida
                descricao:
                  type: string
                  example: Cabo de fibra óptica rompido
      responses:
        '201':
          description: Tipo de falha criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TipoFalha'

  # ==============================================
  # RELATÓRIOS
  # ==============================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token armazenado em HTTPOnly cookie
  responses:
...
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Acesso negado
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensagem de erro
...
    Link:
      type: object
      properties:
        id:
          type: integer
          example: 1
        designador:
          type: string
          example: LINK-001
        tipoServico:
          type: string
          enum: [LINK IP, TRANSPORTE]
          example: LINK IP
        capacidade:
          type: string
          example: 10G
        diaVencimento:
          type: integer
          nullable: true
          minimum: 1
          maximum: 31
          example: 25
        ativo:
          type: boolean
          example: true
        enderecoPontaA:
          type: string
          nullable: true
        enderecoPontaB:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        operadora:
          $ref: '#/components/schemas/Operadora'
        pop:
          $ref: '#/components/schemas/Pop'
        transporte:
          $ref: '#/components/schemas/Transporte'
          nullable: true
    CriarLink:
      type: object
      required:
        - designador
        - operadoraId
        - popId
        - tipoServico
      properties:
        designador:
          type: string
          example: LINK-001
        operadoraId:
          type: integer
          example: 1
        popId:
          type: integer
          example: 1
        transporteId:
          type: integer
          nullable: true
          example: 1
        tipoServico:
          type: string
          enum: [LINK IP, TRANSPORTE]
          example: LINK IP
        capacidade:
          type: string
          example: 10G
        diaVencimento:
          type: integer
          nullable: true
          minimum: 1
          maximum: 31
          example: 25
        enderecoPontaA:
          type: string
        enderecoPontaB:
          type: string
        observacoes:
          type: string
    Transporte:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: TRANSPORTE-001
        fornecedor:
          type: string
          nullable: true
          example: Fornecedor A
        origem:
          type: string
          nullable: true
          example: São Paulo
        destino:
          type: string
          nullable: true
          example: Rio de Janeiro
        ativo:
          type: boolean
          example: true
        observacoes:
          type: string
          nullable: true
        _count:
          type: object
          properties:
            links:
              type: integer
            chamados:
              type: integer
...
    Operadora:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Operadora A
        telefoneSuporte:
          type: string
          nullable: true
          example: '+55 11 1234-5678'
        emailSuporte:
          type: string
          format: email
          nullable: true
          example: suporte@operadora.com
        portalUrl:
          type: string
          format: uri
          nullable: true
          example: https://portal.operadora.com
        slaHoras:
          type: integer
          nullable: true
          example: 4
        ativo:
          type: boolean
          example: true
        observacoes:
          type: string
          nullable: true
    CriarOperadora:
      type: object
      required:
        - nome
      properties:
        nome:
          type: string
          example: Operadora A
        telefoneSuporte:
          type: string
          example: '+55 11 1234-5678'
        emailSuporte:
          type: string
          format: email
          example: suporte@operadora.com
        portalUrl:
          type: string
          format: uri
          example: https://portal.operadora.com
        slaHoras:
          type: integer
          example: 4
        observacoes:
          type: string
    Pop:
      type: object
      properties:
        id:
          type: integer
          example: 1
        codigo:
          type: string
          example: SP01
        nome:
          type: string
          nullable: true
          example: POP São Paulo
        cidade:
          type: string
          example: São Paulo
        estado:
          type: string
          minLength: 2
          maxLength: 2
          example: SP
        endereco:
          type: string
          nullable: true
          example: 'Rua Exemplo, 123'
        ativo:
          type: boolean
          example: true
    CriarPop:
      type: object
      required:
        - codigo
        - cidade
        - estado
      properties:
        codigo:
          type: string
          example: SP01
        nome:
          type: string
          example: POP São Paulo
        cidade:
          type: string
          example: São Paulo
        estado:
          type: string
          minLength: 2
          maxLength: 2
          example: SP
        endereco:
          type: string
          example: 'Rua Exemplo, 123'
...
    TipoFalha:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Fibra Rompida
        descricao:
          type: string
          nullable: true
          example: Cabo de fibra óptica rompido
...
security:
  - cookieAuth: []